ultraman-api/requirements.txt
# Python 3.6 兼容版本依赖
fastapi==0.70.0
uvicorn==0.15.0
gunicorn==20.1.0
sqlalchemy==1.4.27
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
requests==2.27.1
python-multipart==0.0.5